homework/JavaAir/src/main/java/Itinery/Manifest/Manifest.java
package Itinery.Manifest;

import People.Person;

import java.util.ArrayList;

public class Manifest {

        private ArrayList<Person> manifest;

        public Manifest() {this.manifest = new ArrayList<>();
        }

        public void addPersonToManifest(Person person) {
            this.manifest.add(person);
        }

        public Person getPassengerFromManifestIndex(int manifestIndex){
            return manifest.get(manifestIndex);
        }

        public int getTotal() {
            return manifest.size();
        }
}
